[core][clickhouse] update sql tests to compile
